\name{AlleleRetain}
\alias{AlleleRetain}

\title{ Allele Retention, Inbreeding, and Demography }

\description{
Simulates the effect of management or demography on allele retention 
and inbreeding accumulation in bottlenecked populations of animals 
with overlapping generations. 
}

\details{

Typically, the user will run \code{aRetain}, then \code{aRetain.summary} to assess characteristics of the simulated population.  \code{indiv.summary}, \code{pedigree.summary} (requires package \bold{pedigree}),  \code{LRS.summary}, and  \code{agerepro.summary} will provide further output.

}